requirements.txt
python-pptx==0.6.21
requests==2.28.2
beautifulsoup4==4.12.0
